Refactor test.
